replace StateModel using generic hashmap class
