+++ rust/routee-compass-core/src/model/state/mod.rs
@@ -1,9 +1,7 @@
 pub mod custom_feature_format;
 pub mod indexed_state_feature;
-mod internal_state_model_ops;
 pub mod state_error;
 pub mod state_feature;
 pub mod state_model;
-mod state_model_v2;
 pub mod unit_codec_name;
 pub mod update_operation;
