+++ rust/routee-compass-powertrain/src/routee/vehicle/default/ice.rs
@@ -64,7 +64,7 @@ impl VehicleType for ICE {
         let (energy, _energy_unit) = self.best_case_energy(distance)?;
         state_model.add_energy(
             state,
-            ICE::ENERGY_FEATURE_NAME,
+            &ICE::ENERGY_FEATURE_NAME.into(),
             &energy,
             &self
                 .prediction_model_record
@@ -87,7 +87,7 @@ impl VehicleType for ICE {
             .predict(speed, grade, distance)?;
         state_model.add_energy(
             state,
-            ICE::ENERGY_FEATURE_NAME,
+            &ICE::ENERGY_FEATURE_NAME.into(),
             &energy,
             &self
                 .prediction_model_record
